发布到服务器的第一版

+++ se-proj-rear-end/src/main/java/com/evan/seprojrearend/mapper/TeacherMapper.java
@@ -3,6 +3,8 @@ package com.evan.seprojrearend.mapper;
 import com.evan.seprojrearend.po.Teacher;
 import com.evan.seprojrearend.po.TeacherExample;
 import com.evan.seprojrearend.po.TeacherKey;
+
+import java.math.BigDecimal;
 import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
@@ -19,7 +21,7 @@ public interface TeacherMapper {
 
     List<Teacher> selectByExample(TeacherExample example);
 
-    Teacher selectByPrimaryKey(Integer key);
+    Teacher selectByPrimaryKey(BigDecimal key);
 
     int updateByExampleSelective(@Param("record") Teacher record, @Param("example") TeacherExample example);
